+++ test/pow/extension/phoenix/router_test.exs
@@ -1,9 +1,11 @@
 defmodule Pow.Test.Extension.Phoenix.Router.Phoenix.Router do
   use Pow.Extension.Phoenix.Router.Base
 
+  alias Pow.Phoenix.Router
+
   defmacro routes(_config) do
     quote location: :keep do
-      resources "/test", TestController, only: [:new, :create, :edit, :update]
+      Router.pow_resources "/test", TestController, only: [:new, :create, :edit, :update]
     end
   end
 end
@@ -13,6 +15,10 @@ defmodule Pow.Test.Extension.Phoenix.Router do
   use Pow.Phoenix.Router
   use Pow.Extension.Phoenix.Router,
     extensions: [Pow.Test.Extension.Phoenix.Router]
+
+  scope "/", as: "pow_test_extension_phoenix_router" do
+    get "/test/:id/overidden", TestController, :edit
+  end
 
   scope "/" do
     pow_routes()
@@ -51,11 +57,15 @@ defmodule Pow.Extension.Phoenix.RouterTest do
 
   test "has routes" do
     assert unquote(Routes.pow_session_path(@conn, :new)) == "/session/new"
-    assert unquote(Routes.pow_test_extension_phoenix_router_test_path(@conn, :new)) = "/test/new"
+    assert unquote(Routes.pow_test_extension_phoenix_router_test_path(@conn, :new)) == "/test/new"
   end
 
   test "validates no aliases" do
     assert unquote(module_raised_with) =~ "Pow routes should not be defined inside scopes with aliases: Test"
     assert unquote(module_raised_with) =~ "scope \"/\", Test do"
+  end
+
+  test "can override routes" do
+    assert unquote(Routes.pow_test_extension_phoenix_router_test_path(@conn, :edit, 1)) == "/test/1/overidden"
   end
 end